+++ src/index.js
@@ -1,3 +1,0 @@
-// @flow
-export * from './Margin';
-export * from './Padding';